ecommerce/requirements.txt
django>=4.2 
django-stubs>=5.0
djangorestframework
djangorestframework-simplejwt